Started script for making csvs of database
Very bare bones start...currently only works for reaction libraries and
training sets as I think these are the ones people will be using the most.

For training sets the template is which groups the training reaction fall
under. This is to help users decide if they want to add new groups or not.