Change definition of one group and added forbidden group to 1,2shiftS
For testing we need all backbone/end families to have top 'end' groups that are groups. Previously this family had a LogicOr that restricted *1 to be a C atom.
Presumably, it is theoreticaly possible for any R!H to react this way, but we only have data for C.

I have changed that logicOr to a R!H, and created a forbidden group to restrict *1 to C atoms. This should be functionally the same.