Used unadjusted MESMER calculations for OH+phenylacetylene central addition k in training reactions.
Adjusted value was slightly higher, and possibly prevented formation of
naphthalene and indene. Using either the adjusted or unadjusted value has a
negligible effect on the comparison with Goulay's experimental
measurements of OH+phenylacetylene, because at low-T (where experiment
was conducted) terminal addition dominates. Therefore, the actual rate
of central addition is largely uncertain, and adjusting it is not
justified.