Fix some bugs in evaluateKinetics.py
The compareNIST had a major error where units for reverse reaction rates would sometimes be computed incorrectly. This commit fixes that.

It also has a quality of life bug fix where we can now skip families that do not have a NIST depository. Previously we had to make an empty NIST folder for each family.